requirements.txt
fastapi
psycopg-binary
python-multipart
uvicorn[standard]
